pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Amos G. Rocha <amosgr12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
factory-boy = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
